Refactor plugin utils
