+++ pkg/pluginutils/plugin_client_test.go
@@ -1,25 +1,24 @@
-package framework_test
+package pluginutils_test
 
 import (
 	"os"
 	"time"
 
+	"k8s.io/kubectl/pkg/pluginutils"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"k8s.io/kubectl/pkg/framework"
 )
 
 var _ = Describe("InitConfig", func() {
-	var ()
-
 	BeforeEach(func() {
-		os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_KUBECONFIG", "assets/config")
+		os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_KUBECONFIG", "testdata/config")
 	})
 
 	Describe("InitConfig", func() {
 		Context("When nothing is overridden by the calling framework", func() {
 			It("finds and parses the preexisting config", func() {
-				config, err := framework.InitConfig()
+				config, err := pluginutils.InitConfig()
 				Expect(err).NotTo(HaveOccurred())
 
 				Expect(config.Host).To(Equal("https://notreal.com:1234"))
@@ -33,9 +32,9 @@ var _ = Describe("InitConfig", func() {
 				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_AS", "apple")
 				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_AS_GROUP", "[\"banana\",\"cherry\"]")
 
-				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CERTIFICATE_AUTHORITY", "assets/apiserver_ca.crt")
-				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CLIENT_CERTIFICATE", "assets/client.crt")
-				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CLIENT_KEY", "assets/client.key")
+				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CERTIFICATE_AUTHORITY", "testdata/apiserver_ca.crt")
+				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CLIENT_CERTIFICATE", "testdata/client.crt")
+				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CLIENT_KEY", "testdata/client.key")
 
 				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_REQUEST_TIMEOUT", "45s")
 				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_SERVER", "some-other-server.com")
@@ -46,15 +45,15 @@ var _ = Describe("InitConfig", func() {
 				os.Setenv("KUBECTL_PLUGINS_GLOBAL_FLAG_CLUSTER", "")
 			})
 			It("overrides the config settings with the passed in settings", func() {
-				config, err := framework.InitConfig()
+				config, err := pluginutils.InitConfig()
 				Expect(err).NotTo(HaveOccurred())
 
 				Expect(config.Impersonate.UserName).To(Equal("apple"))
 				Expect(config.Impersonate.Groups).Should(ConsistOf("banana", "cherry"))
 
-				Expect(config.CertFile).To(Equal("assets/client.crt"))
-				Expect(config.KeyFile).To(Equal("assets/client.key"))
-				Expect(config.CAFile).To(Equal("assets/apiserver_ca.crt"))
+				Expect(config.CertFile).To(Equal("testdata/client.crt"))
+				Expect(config.KeyFile).To(Equal("testdata/client.key"))
+				Expect(config.CAFile).To(Equal("testdata/apiserver_ca.crt"))
 
 				Expect(config.Timeout).To(Equal(45 * time.Second))
 				Expect(config.ServerName).To(Equal("some-other-server.com"))
