+++ pkg/pluginutils/testdata/apiserver_ca.crt
